+++ docs/changelog.rst
@@ -2,6 +2,14 @@
 Changelog
 =========
 
+
+Changes in -master-
+===================
+* working sessions
+* forms?
+* added support for embedded fields
+
+
 Changes in 0.1
 ===============